# api/openapi.yaml
openapi: 3.1.0
info:
  title: kafSIEM Analyst API
  version: 0.1.0
servers:
  - url: /api/v1
paths:
  /entities/{type}/{id}:
    get:
      operationId: getEntity
      parameters:
        - in: path
          name: type
          required: true
          schema: { type: string }
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Entity profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        default:
          $ref: '#/components/responses/Problem'
  /entities/{type}/{id}/neighborhood:
    get:
      operationId: getEntityNeighborhood
      parameters:
        - in: path
          name: type
          required: true
          schema: { type: string }
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: depth
          schema: { type: integer, minimum: 1, maximum: 3, default: 2 }
        - in: query
          name: types
          schema: { type: string }
        - in: query
          name: window
          schema: { type: string }
      responses:
        '200':
          description: Entity neighborhood
          content:
            application/json:
              schema:
                type: object
                required: [entities, edges]
                properties:
                  entities:
                    type: array
                    items: { $ref: '#/components/schemas/Entity' }
                  edges:
                    type: array
                    items: { $ref: '#/components/schemas/Edge' }
        default:
          $ref: '#/components/responses/Problem'
  /entities/{type}/{id}/provenance:
    get:
      operationId: listEntityProvenance
      parameters:
        - in: path
          name: type
          required: true
          schema: { type: string }
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Provenance chain
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListProvenanceResponse'
        default:
          $ref: '#/components/responses/Problem'
  /entities/{type}/{id}/geometry:
    get:
      operationId: getEntityGeometry
      parameters:
        - in: path
          name: type
          required: true
          schema: { type: string }
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Geometry
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Geometry' }
        default:
          $ref: '#/components/responses/Problem'
  /entities/{type}/{id}/timeline:
    get:
      operationId: listEntityTimeline
      parameters:
        - in: path
          name: type
          required: true
          schema: { type: string }
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: after
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Timeline
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListMessageResponse' }
        default:
          $ref: '#/components/responses/Problem'
  /graph/path:
    get:
      operationId: getGraphPath
      parameters:
        - in: query
          name: src
          required: true
          schema: { type: string }
        - in: query
          name: dst
          required: true
          schema: { type: string }
        - in: query
          name: max
          schema: { type: integer, minimum: 1, maximum: 3, default: 3 }
      responses:
        '200':
          description: Path result
          content:
            application/json:
              schema:
                type: object
                required: [found, edges]
                properties:
                  found: { type: boolean }
                  edges:
                    type: array
                    items: { $ref: '#/components/schemas/Edge' }
        default:
          $ref: '#/components/responses/Problem'
  /map/features:
    get:
      operationId: listMapFeatures
      parameters:
        - in: query
          name: bbox
          required: true
          schema: { type: string, example: '14.40,35.80,14.60,36.00' }
        - in: query
          name: types
          schema: { type: string }
        - in: query
          name: window
          schema: { type: string }
      responses:
        '200':
          description: GeoJSON feature collection
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeatureCollection' }
        default:
          $ref: '#/components/responses/Problem'
...
  /flows:
    get:
      operationId: listFlows
      parameters:
        - in: query
          name: after
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: topic
          schema: { type: string }
        - in: query
          name: sender
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: q
          schema: { type: string }
      responses:
        '200':
          description: Flows
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListFlowResponse' }
        default:
          $ref: '#/components/responses/Problem'
  /flows/{id}:
    get:
      operationId: getFlow
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Flow
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Flow' }
        default:
          $ref: '#/components/responses/Problem'
  /flows/{id}/messages:
    get:
      operationId: listFlowMessages
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: after
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Flow messages
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListMessageResponse' }
        default:
          $ref: '#/components/responses/Problem'
  /flows/{id}/tasks:
    get:
      operationId: listFlowTasks
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Flow tasks
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListTaskResponse' }
        default:
          $ref: '#/components/responses/Problem'
  /flows/{id}/traces:
    get:
      operationId: listFlowTraces
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Flow traces
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListTraceResponse' }
        default:
          $ref: '#/components/responses/Problem'
  /flows/{id}/timeline:
    get:
      operationId: listFlowTimeline
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: after
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Flow timeline
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListMessageResponse' }
...
  /replays:
    get:
      operationId: listReplays
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Replay sessions
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListReplayResponse' }
    post:
      operationId: createReplayRequest
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                topics:
                  type: array
                  items: { type: string }
      responses:
        '202':
          description: Replay request accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReplayRequest' }
        default:
          $ref: '#/components/responses/Problem'
  /search:
    get:
      operationId: searchEntities
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                required: [items, next]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
                  next:
                    type: [string, 'null']
...
components:
  responses:
    Problem:
      description: RFC 9457 problem details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Entity:
      type: object
      required: [id, type, canonical_id, first_seen, last_seen]
      properties:
        id: { type: string }
        type: { type: string }
        canonical_id: { type: string }
        display_name: { type: string }
        first_seen: { type: string, format: date-time }
        last_seen: { type: string, format: date-time }
        attrs:
          type: object
          additionalProperties: true
    Edge:
      type: object
      required: [src_id, dst_id, type, valid_from, weight]
      properties:
        src_id: { type: string }
        dst_id: { type: string }
        type: { type: string }
        valid_from: { type: string, format: date-time }
        valid_to: { type: string, format: date-time }
        evidence_msg: { type: string }
        weight: { type: number }
        attrs:
          type: object
          additionalProperties: true
...
    Geometry:
      type: object
      required: [entity_id, geometry_type, geojson, srid, min_lat, min_lon, max_lat, max_lon, observed_at]
      properties:
        entity_id: { type: string }
        geometry_type: { type: string }
        geojson:
          description: RFC 7946 geometry
        srid: { type: integer }
        min_lat: { type: number }
        min_lon: { type: number }
        max_lat: { type: number }
        max_lon: { type: number }
        z_min: { type: number }
        z_max: { type: number }
        observed_at: { type: string, format: date-time }
        valid_to: { type: string, format: date-time }
...
    Profile:
      type: object
      required: [entity, first_seen, last_seen, edge_counts, top_neighbors]
      properties:
        entity: { $ref: '#/components/schemas/Entity' }
        first_seen: { type: string, format: date-time }
        last_seen: { type: string, format: date-time }
        edge_counts:
          type: object
          additionalProperties: { type: integer }
        top_neighbors:
          type: array
          items: { $ref: '#/components/schemas/Neighbor' }
...
    Flow:
      type: object
      required: [id, topic_count, sender_count, topics, senders, trace_ids, task_ids, first_seen, last_seen, message_count]
      properties:
        id: { type: string }
        topic_count: { type: integer }
        sender_count: { type: integer }
        topics: { type: array, items: { type: string } }
        senders: { type: array, items: { type: string } }
        trace_ids: { type: array, items: { type: string } }
        task_ids: { type: array, items: { type: string } }
        first_seen: { type: string, format: date-time }
        last_seen: { type: string, format: date-time }
        latest_status: { type: string }
        message_count: { type: integer }
        latest_preview: { type: string }
...
    FeatureCollection:
      type: object
      required: [type, features]
      properties:
        type: { type: string, const: FeatureCollection }
        features:
          type: array
          items: { $ref: '#/components/schemas/Feature' }
    ListFlowResponse:
      type: object
      required: [items, next]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Flow' } }
        next: { type: [string, 'null'] }
    ListMessageResponse:
      type: object
      required: [items, next]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Message' } }
        next: { type: [string, 'null'] }
    ListTaskResponse:
      type: object
      required: [items, next]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Task' } }
        next: { type: [string, 'null'] }
    ListTraceResponse:
      type: object
      required: [items, next]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Trace' } }
        next: { type: [string, 'null'] }
    ListReplayResponse:
      type: object
      required: [items, next]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/ReplaySession' } }
        next: { type: [string, 'null'] }
...
    ListProvenanceResponse:
      type: object
      required: [items, next]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Provenance' } }
        next: { type: [string, 'null'] }
...